refactor: rename converter modules
